veganswap/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6
thefuzz==0.20.0
python-Levenshtein==0.23.0
pytest==7.4.4
pytest-asyncio==0.23.3
